Mask_generation/requirements.txt
# Mask Generation Tools - Python Dependencies

# Environment variable management
python-dotenv>=0.19.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Numerical computation
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.4.0

# Optional: for better performance with OpenCV
# opencv-contrib-python>=4.5.0
